商品削除機能（コードレビュー③）

+++ app/controllers/items_controller.rb
@@ -38,9 +38,11 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
+    if current_user.id != @item.user_id 
+      if @item.destroy
       redirect_to root_path
-   end
+      end
+    end  
   end   
 
   private
